feat(view): プラン作成を行う場所をURLクエリから取得する
